requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.2
aiofiles==24.1.0
slowapi==0.1.9
celery==5.4.0
redis==5.0.8
tenacity==9.0.0
python-multipart==0.0.9
